+++ workflows/slack/versions/0.0.2/images/post-to-channel/lib/slack.py
@@ -42,7 +42,6 @@ def main():
 
     try:
         response = client.chat_postMessage(channel=channel, text=message)
-        assert response["message"]["text"] == message
     except SlackApiError as e:
         # You will get a SlackApiError if "ok" is False
         assert e.response["ok"] is False